Change pagination position and remove unnecessary pages
