--- control-station/src/pages/VehiclePage/BoardsPage/BoardsPage.module.scss
@@ -1,11 +1,0 @@
-.boardsPage {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    flex: 1;
-    gap: 1rem;
-
-    > * {
-        min-width: fit-content;
-    }
-}
