+++ control-station/src/pages/VehiclePage/Data1Page/BoardsPage.module.scss
@@ -1,0 +1,17 @@
+.container {
+    display: flex;
+    flex-flow: row;
+    gap: 1rem;
+}
+
+.boardsPage {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    flex: 1;
+    gap: 1rem;
+
+    > * {
+        min-width: fit-content;
+    }
+}
