+++ control-station/src/pages/VehiclePage/Data2Page/BoardsPage.module.scss
@@ -1,0 +1,11 @@
+.boardsPage {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    flex: 1;
+    gap: 1rem;
+
+    > * {
+        min-width: fit-content;
+    }
+}
