+++ control-station/src/pages/VehiclePage/VehiclePage.module.scss
@@ -1,0 +1,7 @@
+.pagination_position {
+    position: fixed;
+    right: 50%;
+    bottom: 1rem;
+
+    width: fit-content;
+}
