+++ control-station/src/pages/VehiclePage/vehicleRoute.tsx
@@ -1,16 +1,16 @@
-import { BoardsPage1 } from "./BoardsPage/BoardsPage1";
-import { BoardsPage2 } from "./BoardsPage/BoardsPage2";
-import { ControlPage } from "./ControlPage/ControlPage";
-import { VehiclePage } from "./VehiclePage";
-import { Navigate } from "react-router-dom";
+import { ControlPage } from './ControlPage/ControlPage';
+import { Data1Page } from './Data1Page/Data1Page';
+import { Data2Page } from './Data2Page/Data2Page';
+import { VehiclePage } from './VehiclePage';
+import { Navigate } from 'react-router-dom';
 
 export const vehicleRoute = {
-    path: "/vehicle",
+    path: '/vehicle',
     element: <VehiclePage />,
     children: [
-        { path: "", element: <Navigate to={"first"} /> },
-        { path: "first", element: <BoardsPage1 /> },
-        { path: "second", element: <BoardsPage2 /> },
-        { path: "thirst", element: <ControlPage /> },
+        { path: '', element: <Navigate to={'data-1'} /> },
+        { path: 'data-1', element: <Data1Page /> },
+        { path: 'data-2', element: <Data2Page /> },
+        { path: 'controls', element: <ControlPage /> },
     ],
 };
